+++ lib/src/bones_api_server.dart
@@ -21,6 +21,7 @@ import 'bones_api_config.dart';
 import 'bones_api_extension.dart';
 import 'bones_api_hotreload.dart';
 import 'bones_api_logging.dart';
+import 'bones_api_session.dart';
 import 'bones_api_types.dart';
 import 'bones_api_utils_httpclient.dart';
 import 'bones_api_utils_json.dart';